Sprint 08

+++ stories/src/use-cases/delete-story.use-case.ts
@@ -2,7 +2,11 @@ import { StoryRepository } from "../repositories/story.repository";
 export class DeleteStoryUseCase {
     constructor(private readonly repository: StoryRepository) {}
 
-    execute(storyUuid: string): Promise<void> {
+    public async execute(storyUuid: string): Promise<void> {
+        const uuidExists = await this.repository.doesUuidExists(storyUuid);
+        if (!uuidExists) {
+            throw new Error('["story uuid not found"]');
+        }
         return this.repository.delete(storyUuid);
     }
 }
